Progress toward Producer Algebra and config package reorg
